test: 각 컴포넌트의 테스트 코드 추가
